web: move router to elements
