+++ web/src/main.ts
@@ -19,11 +19,10 @@ import "./elements/CodeMirror";
 import "./elements/Messages";
 import "./elements/Spinner";
 import "./elements/Tabs";
+import "./elements/router/RouterOutlet";
 
 import "./pages/generic/FlowShellCard";
 import "./pages/generic/SiteShell";
-
-import "./pages/router/RouterOutlet";
 
 import "./pages/admin-overview/AdminOverviewPage";
 import "./pages/admin-overview/TopApplicationsTable";
